requirements.txt
et-xmlfile==1.1.0
numpy==1.21.2
openpyxl==3.0.7
pandas==1.3.2
pymongo==3.12.0
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
six==1.16.0
